finish doing 1 HW

+++ Practise/BottleOfWater.java
@@ -2,7 +2,7 @@ package Practise;
 
 public class BottleOfWater extends Product {
 
-    public BottleOfWater(String name, double cost) {
+    public BottleOfWater(String name, int cost) {
         super(name, cost);
     }
 }
