+++ Practise/Fanta.java
@@ -1,7 +1,7 @@
 package Practise;
 
 public class Fanta extends Product{
-    public Fanta(String name, double cost) {
+    public Fanta(String name, int cost) {
         super(name, cost);
     }
 }
