+++ Practise/Snikers.java
@@ -1,7 +1,7 @@
 package Practise;
 
 public class Snikers extends Product {
-    public Snikers(String name, double cost) {
+    public Snikers(String name, int cost) {
         super(name, cost);
     }
 }
